MergeTreeFileFormat trait
