+++ backends-clickhouse/src/main/delta-20/org/apache/spark/sql/execution/datasources/v2/clickhouse/source/DeltaMergeTreeFileFormat.scala
@@ -16,23 +16,12 @@
  */
 package org.apache.spark.sql.execution.datasources.v2.clickhouse.source
 
-import org.apache.gluten.execution.datasource.GlutenFormatFactory
-
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.delta.DeltaParquetFileFormat
+import org.apache.spark.sql.delta.{DeltaParquetFileFormat, MergeTreeFileFormat}
 import org.apache.spark.sql.delta.actions.Metadata
-import org.apache.spark.sql.execution.datasources.{OutputWriter, OutputWriterFactory}
-import org.apache.spark.sql.execution.datasources.mergetree.DeltaMetaReader
-import org.apache.spark.sql.types.StructType
 
-import org.apache.hadoop.mapreduce.{Job, TaskAttemptContext}
-
-class DeltaMergeTreeFileFormat(metadata: Metadata)
-  extends DeltaParquetFileFormat(metadata.columnMappingMode, metadata.schema) {
-
-  override def shortName(): String = "mergetree"
-
-  override def toString: String = "MergeTree"
+class DeltaMergeTreeFileFormat(val metadata: Metadata)
+  extends DeltaParquetFileFormat(metadata.columnMappingMode, metadata.schema)
+  with MergeTreeFileFormat {
 
   override def equals(other: Any): Boolean = {
     other match {
@@ -43,33 +32,4 @@ class DeltaMergeTreeFileFormat(metadata: Metadata)
   }
 
   override def hashCode(): Int = getClass.getCanonicalName.hashCode()
-
-  override def prepareWrite(
-      sparkSession: SparkSession,
-      job: Job,
-      options: Map[String, String],
-      dataSchema: StructType): OutputWriterFactory = {
-    // pass compression to job conf so that the file extension can be aware of it.
-    val conf = job.getConfiguration
-
-    val nativeConf = GlutenFormatFactory(shortName()).nativeConf(options, "")
-
-    @transient val deltaMetaReader = DeltaMetaReader(metadata)
-    deltaMetaReader.storageConf.foreach { case (k, v) => conf.set(k, v) }
-
-    new OutputWriterFactory {
-
-      /** no extension for MergeTree */
-      override def getFileExtension(context: TaskAttemptContext): String = ""
-
-      override def newInstance(
-          path: String,
-          dataSchema: StructType,
-          context: TaskAttemptContext): OutputWriter = {
-
-        GlutenFormatFactory(shortName())
-          .createOutputWriter(path, metadata.schema, context, nativeConf)
-      }
-    }
-  }
 }
