+++ backends-clickhouse/src/main/delta-32/org/apache/spark/sql/execution/datasources/v2/clickhouse/source/DeltaMergeTreeFileFormat.scala
@@ -16,24 +16,13 @@
  */
 package org.apache.spark.sql.execution.datasources.v2.clickhouse.source
 
-import org.apache.gluten.execution.datasource.GlutenFormatFactory
-
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.delta.DeltaParquetFileFormat
+import org.apache.spark.sql.delta.{DeltaParquetFileFormat, MergeTreeFileFormat}
 import org.apache.spark.sql.delta.actions.{Metadata, Protocol}
-import org.apache.spark.sql.execution.datasources.{OutputWriter, OutputWriterFactory}
-import org.apache.spark.sql.execution.datasources.mergetree.DeltaMetaReader
-import org.apache.spark.sql.types.StructType
-
-import org.apache.hadoop.mapreduce.{Job, TaskAttemptContext}
 
 @SuppressWarnings(Array("io.github.zhztheplayer.scalawarts.InheritFromCaseClass"))
 class DeltaMergeTreeFileFormat(protocol: Protocol, metadata: Metadata)
-  extends DeltaParquetFileFormat(protocol, metadata) {
-
-  override def shortName(): String = "mergetree"
-
-  override def toString: String = "MergeTree"
+  extends DeltaParquetFileFormat(protocol, metadata)
+  with MergeTreeFileFormat {
 
   override def equals(other: Any): Boolean = {
     other match {
@@ -46,33 +35,4 @@ class DeltaMergeTreeFileFormat(protocol: Protocol, metadata: Metadata)
   }
 
   override def hashCode(): Int = getClass.getCanonicalName.hashCode()
-
-  override def prepareWrite(
-      sparkSession: SparkSession,
-      job: Job,
-      options: Map[String, String],
-      dataSchema: StructType): OutputWriterFactory = {
-    // pass compression to job conf so that the file extension can be aware of it.
-    val conf = job.getConfiguration
-
-    val nativeConf = GlutenFormatFactory(shortName()).nativeConf(options, "")
-
-    @transient val deltaMetaReader = DeltaMetaReader(metadata)
-    deltaMetaReader.storageConf.foreach { case (k, v) => conf.set(k, v) }
-
-    new OutputWriterFactory {
-      override def getFileExtension(context: TaskAttemptContext): String = {
-        ".mergetree"
-      }
-
-      override def newInstance(
-          path: String,
-          dataSchema: StructType,
-          context: TaskAttemptContext): OutputWriter = {
-
-        GlutenFormatFactory(shortName())
-          .createOutputWriter(path, metadata.schema, context, nativeConf)
-      }
-    }
-  }
 }
